update issue: 1、modifity some file;
