+++ app/build.gradle
@@ -24,7 +24,7 @@ android {
     }
     buildTypes {
         release {
-            minifyEnabled true
+            minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
             buildConfigField "String", "So_Library_name", "\"connrotCmake-lib\""    //connrot-lib
         }
