+++ app/src/main/java/com/example/ShikeApplication/carmera/CameraSurfaceView.java
@@ -14,11 +14,8 @@ import android.util.Log;
 import android.view.SurfaceHolder;
 import android.view.SurfaceView;
 
-import com.example.ShikeApplication.carmera.utils.BitmapUtils;
-import com.example.ShikeApplication.carmera.utils.MetricsUtil;
-
-import java.lang.reflect.Array;
-import java.util.ArrayList;
+import com.example.ShikeApplication.utils.BitmapUtils;
+import com.example.ShikeApplication.utils.MetricsUtil;
 
 public class CameraSurfaceView extends SurfaceView {
