+++ app/src/main/java/com/example/ShikeApplication/mediacodec/MediaUtil.java
@@ -1,0 +1,6 @@
+package com.example.ShikeApplication.mediacodec;
+
+public class MediaUtil {
+	private final static String TAG = "MediaUtil";
+
+}
