+++ app/src/main/java/com/example/ShikeApplication/utils/MetricsUtil.java
@@ -1,4 +1,4 @@
-package com.example.ShikeApplication.carmera.utils;
+package com.example.ShikeApplication.utils;
 
 import java.util.regex.Pattern;
